WiFiHAL: Remove unused param warnings in WifiEvent class
Compiler throws a warning when it finds unused parameters in
WifiEvent class. The parameters showing as unused are member
funtion arguments and the member functions are going to be
overriden in derived classes. It's not good to change the
prototype of the functions to resolve these errors.
So introduce a dummy define UNUSED() to use the parameter and
mute the warning by calling the same.

Change-Id: I0abe95e9490a1874968a506e38ccdac25edd9dd2
CRs-Fixed: 1086199